Convert character and game object collision scripts to engine classes
